Split acid-state-internal package out of acid-state
