+++ acid-state-internal/acid-state-internal.cabal
@@ -1,0 +1,59 @@
+Name:                acid-state-internal
+Version:             0.15.0
+Synopsis:            Add ACID guarantees to any serializable Haskell data structure.
+Description:         Use regular Haskell data structures as your database and get stronger ACID guarantees than most RDBMS offer.
+Homepage:            https://github.com/acid-state/acid-state
+License:             PublicDomain
+Author:              David Himmelstrup
+Maintainer:          Lemmih <lemmih@gmail.com>
+-- Copyright:
+Category:            Database
+Build-type:          Simple
+Cabal-version:       >=1.10
+tested-with:         GHC == 7.6.3, GHC == 7.8.4, GHC == 7.10.3, GHC == 8.0.2, GHC == 8.2.2, GHC == 8.4.3
+
+Source-repository head
+  type:          git
+  location:      https://github.com/acid-state/acid-state
+
+Library
+  Exposed-Modules:     Data.Acid.Internal.Archive
+                       Data.Acid.Internal.Core
+                       Data.Acid.Internal.Local
+                       Data.Acid.Internal.TemplateHaskell
+                       Data.Acid.Abstract
+                       Data.Acid.Common
+                       Data.Acid.CRC
+                       Data.Acid.Log
+                       Data.Acid.Memory
+                       Data.Acid.Memory.Pure
+
+  Other-modules:       Data.Acid.Internal.FileIO
+                       Paths_acid_state_internal
+
+  Build-depends:       array,
+                       base >= 4.5.1.0 && < 5,
+                       bytestring >= 0.10.2,
+                       containers,
+                       stm >= 2.4,
+                       directory,
+                       filelock,
+                       filepath,
+                       mtl,
+                       template-haskell,
+                       th-expand-syns
+
+  if os(windows)
+     Build-depends:    Win32
+  else
+     Build-depends:    unix
+
+  Hs-Source-Dirs:      src/
+
+  if os(windows)
+     Hs-Source-Dirs:   src-win32/
+  else
+     Hs-Source-Dirs:   src-unix/
+
+  default-language:    Haskell2010
+  GHC-Options:         -fwarn-unused-imports -fwarn-unused-binds
