+++ acid-state-internal/src/Data/Acid/CRC.hs
